cloud-organizer/pyproject.toml
[project]
name = "cloud-organizer"
version = "0.1.0"
description = "Index, tag, and sync files across multiple cloud services with Notion integration"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0",
    "flask>=3.0",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "msal>=1.24",
    "requests>=2.31",
    "dropbox>=12.0",
    "notion-client>=2.0",
    "google-generativeai>=0.5",
    "pyyaml>=6.0",
]

[project.scripts]
cloud-organizer = "cloud_organizer.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
